+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Emission
 
 - [Messaging] Adds inbox zero state - maxim
+- [Messaging] Relay support in Conversation container - sarah
 - [Messaging] Adds a FlatList for Messages in Conversations - sarah
 - [dev] Applies prettier, and adds it to the webhooks + extension settings - orta
 - [dev] Adds stories for new messaging components - maxim